refactor empty-bin command and wrote javadoc comments
